+++ config/splat.us.stnz0.yaml
@@ -84,7 +84,7 @@ segments:
       - [0x305B8, rodata]
       - [0x305E8, .rodata, 33FCC] # EntityBossRoomBlock
       - [0x305FC, .rodata, 33FCC] # EntitySlogra
-      - [0x30618, rodata] # EntityGaibon .rodata, 33FCC
+      - [0x30618, .rodata, gaibon] # EntityGaibon .rodata, 33FCC
       - [0x3063C, .rodata, 36DE4] # func_801B7034
       - [0x3064C, rodata]
       - [0x30728, .rodata, 39908] # EntityRedDoor
@@ -105,6 +105,7 @@ segments:
       - [0x30934, rodata]
       - [0x30958, c]
       - [0x33FCC, c] # Slogra & Gaibon Boss
+      - [0x35778, c, gaibon]
       - [0x36DE4, c]
       - [0x394D4, c, random]
       - [0x39504, c, update]
