+++ config/symbols.us.stnz0.txt
@@ -68,7 +68,6 @@ EntitySmallGaibonProjectile = 0x801B6AAC;
 EntityLargeGaibonProjectile = 0x801B6BBC;
 EntityElevator = 0x801B7034;
 EntityMariaCutscene = 0x801B7D58;
-Random = 0x801B94D4;
 Update = 0x801B9504;
 UpdateStageEntities = 0x801B9800;
 HitDetection = 0x801B9908;
@@ -84,7 +83,6 @@ GetSideToPlayer = 0x801BCC5C;
 MoveEntity = 0x801BCCA0;
 AllocEntity = 0x801BD15C;
 SetStep = 0x801BD52C;
-SetSubStep = 0x801BD54C;
 InitializeEntity = 0x801BD5FC;
 EntityDummy = 0x801BD6F8;
 GetPlayerCollisionWith = 0x801BD9A0;
