+++ src/st/nz0/33FCC.c
@@ -5,14 +5,7 @@
 
 #include "nz0.h"
 
-#define BOSS_FLAG_DOORS_CLOSED (1 << 0)
-#define BOSS_FLAG_DOORS_OPEN (1 << 1)
-#define BOSS_FLAG_FIGHT_BEGIN (1 << 2)
-#define BOSS_FLAG_GAIBON_DEAD (1 << 3)
-#define BOSS_FLAG_SLOGRA_DEAD (1 << 4)
-
 #define GAIBON self[8]
-#define SLOGRA self[-8]
 
 typedef enum {
     SLOGRA_INIT,
@@ -56,71 +49,6 @@ typedef enum {
     SLOGRA_DYING_EXPLODING,
     SLOGRA_DYING_END,
 } SlograDyingSubSteps;
-
-typedef enum {
-    GAIBON_INIT,
-    GAIBON_IDLE = 2,
-    GAIBON_FLY_TOWARDS_PLAYER,
-    GAIBON_FLY_SHOOT_FIREBALLS,
-    GAIBON_LANDING_AFTER_SHOOTING,
-    GAIBON_SHOOT_FROM_GROUND,
-    GAIBON_FLY_SHOOT_BIG_FIREBALL = 8,
-    GAIBON_PICKUP_SLOGRA,
-    GAIBON_NEAR_DEATH = 15,
-    GAIBON_DYING,
-    GAIBON_DEBUG = 255,
-} GaibonSteps;
-
-typedef enum {
-    GAIBON_FLY_TOWARDS_PLAYER_BEGIN,
-    GAIBON_FLY_TOWARDS_PLAYER_MOVEMENT,
-    GAIBON_FLY_TOWARDS_PLAYER_END,
-} GaibonFlyTowardsPlayerSubSteps;
-
-typedef enum {
-    GAIBON_FLY_SHOOT_FIREBALLS_BEGIN,
-    GAIBON_FLY_SHOOT_FIREBALLS_MOVING_SHOOTING,
-    GAIBON_FLY_SHOOT_FIREBALLS_END,
-} GaibonFlyShootFirewalls_SubSteps;
-
-typedef enum {
-    GAIBON_LANDING_AFTER_SHOOTING_SETUP,
-    GAIBON_FALLING_WITHOUT_MAP_COLLISION,
-    GAIBON_FALLING_WITH_MAP_COLLISION,
-    GAIBON_FALLING_ON_GROUND,
-} GaibonLandingAfterShootingSubSteps;
-
-typedef enum {
-    GAIBON_SHOOT_FROM_GROUND_FACE_PLAYER,
-    GAIBON_SHOOT_FROM_GROUND_FACE_SETUP,
-    GAIBON_SHOOT_FROM_GROUND_FACE_SHOOTING,
-} GaibonShootFromGroundSubSteps;
-
-typedef enum {
-    GAIBON_FLY_SHOOT_BIG_FIREBALL_SETUP,
-    GAIBON_FLY_SHOOT_BIG_FIREBALL_SHOOTING,
-    GAIBON_FLY_SHOOT_BIG_FIREBALL_END,
-} GaibonFlyShootBigFirewallSubSteps;
-
-typedef enum {
-    GAIBON_PICKUP_SLOGRA_SETUP,
-    GAIBON_PICKUP_SLOGRA_MOVING,
-    GAIBON_PICKUP_SLOGRA_ASCENDING,
-    GAIBON_PICKUP_SLOGRA_AIMING,
-    GAIBON_PICKUP_SLOGRA_RELEASE,
-} GaibonPickupSlograSubSteps;
-
-typedef enum {
-    GAIBON_NEAR_DEATH_SETUP,
-    GAIBON_NEAR_DEATH_FLOOR_HIT_WAIT,
-    GAIBON_NEAR_DEATH_FLOOR_LANDING,
-    GAIBON_NEAR_DEATH_TRANSFORM,
-} GaibonNearDeathSubSteps;
-
-typedef enum {
-    GAIBON_DYING_REACT,
-    GAIBON_DYING_TURN_INTO_BONES,
-} GaibonDyingSubSteps;
 
 // DECOMP_ME_WIP EntityCloseBossRoom https://decomp.me/scratch/bqgN9 95.04 %
 // figuring out D_80181014 struct might help
@@ -676,675 +604,6 @@ void EntitySlograSpearProjectile(Entity* self) {
     case 2:
         MoveEntity();
         AnimateEntity(D_80181170, self);
-        break;
-    }
-}
-
-// Confirmed equivalent behaviour in game.
-// Mostly regalloc, instruction reordering
-// https://decomp.me/scratch/QGtpG
-#ifndef NON_MATCHING
-INCLUDE_ASM("st/nz0/nonmatchings/33FCC", EntityGaibon);
-#else
-void EntityGaibon(Entity* self) {
-    s32 slograGaibonDistX;
-    s32 slograGaibonDistY;
-    Collider collider;
-    Entity* newEntity;
-    s32 animStatus;
-    s32 shootSpeed;
-    s16 hitPoints;
-    s32 maxSpeed;
-    s8* hitbox;
-    s16 angle;
-    s32 speed;
-    s32 step; // !FAKE
-    s32 var_s3;
-
-    if (self->step != GAIBON_INIT && self->ext.GS_Props.nearDeath == 0) {
-        hitPoints = g_api.enemyDefs[254].hitPoints;
-        if (self->hitPoints < hitPoints / 2) {
-            self->ext.GS_Props.grabedAscending = 0;
-            self->ext.GS_Props.nearDeath = 1;
-            SetStep(GAIBON_NEAR_DEATH);
-        }
-    }
-
-    if (self->flags & FLAG_DEAD && self->step < GAIBON_NEAR_DEATH) {
-        self->ext.GS_Props.grabedAscending = 0;
-        self->unk3C = 0;
-        SetStep(GAIBON_NEAR_DEATH);
-    } else if (SLOGRA.ext.GS_Props.pickupFlag != 0) {
-        if (self->step < GAIBON_LANDING_AFTER_SHOOTING) {
-            SetStep(GAIBON_PICKUP_SLOGRA);
-        }
-    }
-
-    switch (self->step) {
-    case GAIBON_INIT:
-        InitializeEntity(D_80180D30);
-        CreateEntityFromCurrentEntity(0x44, &self[1]);
-        self[1].zPriority = self->zPriority + 4;
-        SetStep(GAIBON_IDLE);
-        break;
-
-    case GAIBON_IDLE:
-        AnimateEntity(D_80181250, self);
-        if (self->animFrameIdx == 1 && self->animFrameDuration == 0) {
-            func_801C29B0(NA_SE_EN_GAIBON_FLAP_WINGS);
-        }
-        if (self->unk48 != 0) {
-            g_BossFlag |= BOSS_FLAG_FIGHT_BEGIN;
-        }
-        if (g_BossFlag & BOSS_FLAG_FIGHT_BEGIN) {
-            SetStep(GAIBON_FLY_TOWARDS_PLAYER);
-        }
-        break;
-
-    case GAIBON_FLY_TOWARDS_PLAYER:
-        switch (self->step_s) {
-        case GAIBON_FLY_TOWARDS_PLAYER_BEGIN:
-            self->facingLeft = (GetSideToPlayer() & 1) ^ 1;
-            self->ext.GS_Props.angle =
-                ratan2((PLAYER.posY.i.hi - 32) - self->posY.i.hi,
-                       PLAYER.posX.i.hi - self->posX.i.hi);
-            self->ext.GS_Props.speed = 0;
-            self->ext.GS_Props.timer = 96;
-            if (self->ext.GS_Props.nearDeath != 0) {
-                self->ext.GS_Props.timer = 48;
-            }
-            self->step_s++;
-
-        case GAIBON_FLY_TOWARDS_PLAYER_MOVEMENT:
-            if (self->ext.GS_Props.nearDeath != 0) {
-                maxSpeed = 0x40000;
-            } else {
-                maxSpeed = 0x20000;
-            }
-            self->ext.GS_Props.speed += 0xA00;
-            if (self->ext.GS_Props.speed >= maxSpeed) {
-                self->ext.GS_Props.speed = maxSpeed;
-            }
-            speed = self->ext.GS_Props.speed;
-            self->velocityX = (speed * rcos(self->ext.GS_Props.angle)) >> 0xC;
-            self->velocityY = (speed * rsin(self->ext.GS_Props.angle)) >> 0xC;
-            MoveEntity();
-            AnimateEntity(D_80181250, self);
-            if (self->animFrameIdx == 1 && self->animFrameDuration == 0) {
-                func_801C29B0(NA_SE_EN_GAIBON_FLAP_WINGS);
-            }
-            self->facingLeft = (GetSideToPlayer() & 1) ^ 1;
-            self->ext.GS_Props.timer--;
-            if (self->ext.GS_Props.timer == 0) {
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_FLY_TOWARDS_PLAYER_END:
-            MoveEntity();
-            self->velocityX -= self->velocityX / 32;
-            self->velocityY -= self->velocityY / 32;
-            if (AnimateEntity(D_80181264, self) == 0) {
-                SetStep(GAIBON_FLY_SHOOT_FIREBALLS);
-            }
-            if (self->animFrameIdx == 1 && self->animFrameDuration == 0) {
-                func_801C29B0(NA_SE_EN_GAIBON_FLAP_WINGS);
-            }
-        }
-        break;
-
-    // flying up while shooting down at an angle
-    case GAIBON_FLY_SHOOT_FIREBALLS:
-        switch (self->step_s) {
-        case GAIBON_FLY_SHOOT_FIREBALLS_BEGIN:
-            self->facingLeft = (GetSideToPlayer() & 1) ^ 1;
-            newEntity = &PLAYER;
-            var_s3 = newEntity->posX.i.hi;
-            if (GetSideToPlayer() & 1) {
-                var_s3 += 0x60;
-            } else {
-                var_s3 -= 0x60;
-            }
-            self->ext.GS_Props.angle =
-                ratan2(newEntity->posY.i.hi - 128 - self->posY.i.hi,
-                       var_s3 - self->posX.i.hi);
-            self->ext.GS_Props.speed = 0;
-            self->ext.GS_Props.timer = 80;
-            if (self->ext.GS_Props.nearDeath != 0) {
-                self->ext.GS_Props.timer = 40;
-            }
-            self->step_s++;
-
-        case GAIBON_FLY_SHOOT_FIREBALLS_MOVING_SHOOTING:
-            if (self->ext.GS_Props.nearDeath != 0) {
-                maxSpeed = 0x40000;
-            } else {
-                maxSpeed = 0x20000;
-            }
-            self->ext.GS_Props.speed += 0xA00;
-            if (self->ext.GS_Props.speed >= maxSpeed) {
-                self->ext.GS_Props.speed = maxSpeed;
-            }
-            speed = self->ext.GS_Props.speed;
-            self->velocityX = (speed * rcos(self->ext.GS_Props.angle)) >> 0xC;
-            self->velocityY = (speed * rsin(self->ext.GS_Props.angle)) >> 0xC;
-            MoveEntity();
-            AnimateEntity(D_80181298, self);
-            if (self->animFrameIdx == 1 && self->animFrameDuration == 0) {
-                func_801C29B0(NA_SE_EN_GAIBON_FLAP_WINGS);
-            }
-
-            if (self->ext.GS_Props.nearDeath != 0) {
-                shootSpeed = 7;
-            } else {
-                shootSpeed = 15;
-            }
-            if (self->ext.GS_Props.timer & shootSpeed == 0) {
-                newEntity = AllocEntity(g_Entities + 160, g_Entities + 192);
-                if (newEntity != NULL) {
-                    CreateEntityFromEntity(
-                        E_GAIBON_SMALL_FIREBALL, self, newEntity);
-                    func_801C29B0(NA_SE_EN_GAIBON_SMALL_FIREBALL);
-                    newEntity->posY.i.hi -= 2;
-                    if (self->facingLeft != 0) {
-                        newEntity->unk1E = 0x220;
-                        newEntity->posX.i.hi += 12;
-                    } else {
-                        newEntity->unk1E = 0x5E0;
-                        newEntity->posX.i.hi -= 12;
-                    }
-                }
-            }
-            newEntity->zPriority = self->zPriority + 1;
-            self->ext.GS_Props.timer--;
-            if (self->ext.GS_Props.timer == 0) {
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_FLY_SHOOT_FIREBALLS_END:
-            MoveEntity();
-            self->velocityX -= self->velocityX / 32;
-            self->velocityY -= self->velocityY / 32;
-            if (AnimateEntity(D_80181264, self) == 0) {
-                SetStep(GAIBON_LANDING_AFTER_SHOOTING);
-                if (self->ext.GS_Props.nearDeath != 0) {
-                    SetStep(GAIBON_FLY_SHOOT_BIG_FIREBALL);
-                }
-            }
-            if (self->animFrameIdx == 1 && self->animFrameDuration == 0) {
-                func_801C29B0(NA_SE_EN_GAIBON_FLAP_WINGS);
-            }
-        }
-        break;
-
-    case GAIBON_LANDING_AFTER_SHOOTING:
-        switch (self->step_s) {
-        case GAIBON_LANDING_AFTER_SHOOTING_SETUP:
-            self->animCurFrame = 9;
-            self->velocityX = 0;
-            self->velocityY = 0;
-            self->step_s++;
-
-        case GAIBON_FALLING_WITHOUT_MAP_COLLISION:
-            MoveEntity();
-            self->velocityY += FIX(0.09375);
-            if (PLAYER.posY.i.hi - self->posY.i.hi < 48) {
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_FALLING_WITH_MAP_COLLISION:
-            MoveEntity();
-            self->velocityY += FIX(0.09375);
-            g_api.CheckCollision(
-                self->posX.i.hi, self->posY.i.hi + 28, &collider, 0);
-            if (collider.effects & EFFECT_SOLID) {
-                self->posY.i.hi = self->posY.i.hi + *(u16*)&collider.unk18;
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_FALLING_ON_GROUND:
-            if (AnimateEntity(D_801812C0, self) == 0) {
-                SetStep(GAIBON_SHOOT_FROM_GROUND);
-            }
-        }
-        break;
-
-    case GAIBON_SHOOT_FROM_GROUND:
-        switch (self->step_s) {
-        case GAIBON_SHOOT_FROM_GROUND_FACE_PLAYER:
-            self->facingLeft = (GetSideToPlayer() & 1) ^ 1;
-            self->step_s++;
-
-        case GAIBON_SHOOT_FROM_GROUND_FACE_SETUP:
-            if (AnimateEntity(D_801812CC, self) == 0) {
-                self->ext.GS_Props.timer = 64;
-                if (self->ext.GS_Props.nearDeath != 0) {
-                    self->ext.GS_Props.timer = 128;
-                }
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_SHOOT_FROM_GROUND_FACE_SHOOTING:
-            if (!(self->ext.GS_Props.timer % 16)) {
-                newEntity = AllocEntity(g_Entities + 160, g_Entities + 192);
-                if (newEntity != NULL) {
-                    if (self->ext.GS_Props.nearDeath == 0) {
-                        CreateEntityFromEntity(
-                            E_GAIBON_SMALL_FIREBALL, self, newEntity);
-                        func_801C29B0(NA_SE_EN_GAIBON_SMALL_FIREBALL);
-                    } else {
-                        CreateEntityFromEntity(
-                            E_GAIBON_BIG_FIREBALL, self, newEntity);
-                        func_801C29B0(NA_SE_EN_GAIBON_BIG_FIREBALL);
-                    }
-                    newEntity->posY.i.hi -= 6;
-                    if (self->facingLeft != 0) {
-                        newEntity->unk1E = 0;
-                        newEntity->posX.i.hi += 16;
-                    } else {
-                        newEntity->unk1E = 0x800;
-                        newEntity->posX.i.hi -= 16;
-                    }
-                    newEntity->zPriority = self->zPriority + 1;
-                }
-            }
-            self->ext.GS_Props.timer--;
-            if (self->ext.GS_Props.timer == 0) {
-                SetStep(GAIBON_FLY_TOWARDS_PLAYER);
-            }
-        }
-        break;
-
-    case GAIBON_FLY_SHOOT_BIG_FIREBALL:
-        switch (self->step_s) {
-        case GAIBON_FLY_SHOOT_BIG_FIREBALL_SETUP:
-            if (AnimateEntity(D_801812DC, self) == 0) {
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_FLY_SHOOT_BIG_FIREBALL_SHOOTING:
-            if (AllocEntity(g_Entities + 160, g_Entities + 192) != 0) {
-                newEntity = AllocEntity(g_Entities + 160, g_Entities + 192);
-                if (newEntity != NULL) {
-                    CreateEntityFromEntity(
-                        E_GAIBON_BIG_FIREBALL, self, newEntity);
-                    func_801C29B0(NA_SE_EN_GAIBON_BIG_FIREBALL);
-                    newEntity->posY.i.hi -= 2;
-                    if (self->facingLeft != 0) {
-                        newEntity->unk1E = 0x220;
-                        newEntity->posX.i.hi += 12;
-                    } else {
-                        newEntity->unk1E = 0x5E0;
-                        newEntity->posX.i.hi -= 12;
-                    }
-                    newEntity->zPriority = self->zPriority + 1;
-                }
-            }
-            self->velocityY = FIX(-2);
-            if (self->facingLeft != 0) {
-                self->velocityX = FIX(-2);
-            } else {
-                self->velocityX = FIX(2);
-            }
-            self->ext.GS_Props.timer = 32;
-            self->step_s++;
-
-        case GAIBON_FLY_SHOOT_BIG_FIREBALL_END:
-            MoveEntity();
-            self->velocityX -= self->velocityX / 16;
-            self->velocityY -= self->velocityY / 16;
-            self->ext.GS_Props.timer--;
-            if (self->ext.GS_Props.timer == 0) {
-                SetStep(GAIBON_LANDING_AFTER_SHOOTING);
-            }
-        }
-        break;
-
-    case GAIBON_PICKUP_SLOGRA:
-        switch (self->step_s) {
-        case GAIBON_PICKUP_SLOGRA_SETUP:
-            if (SLOGRA.posX.i.hi - self->posX.i.hi > 0) {
-                self->facingLeft = 1;
-            } else {
-                self->facingLeft = 0;
-            }
-            self->ext.GS_Props.speed = 0;
-            self->step_s++;
-
-        case GAIBON_PICKUP_SLOGRA_MOVING:
-            slograGaibonDistX = SLOGRA.posX.i.hi - self->posX.i.hi;
-            slograGaibonDistY = (SLOGRA.posY.i.hi - 28) - self->posY.i.hi;
-            angle = ratan2(slograGaibonDistY, slograGaibonDistX);
-            self->ext.GS_Props.speed += 0x8000;
-            if (self->ext.GS_Props.speed >= 0x38000) {
-                self->ext.GS_Props.speed = 0x38000;
-            }
-            speed = self->ext.GS_Props.speed;
-            self->velocityX = (speed * rcos(angle)) >> 0xC;
-            self->velocityY = (speed * rsin(angle)) >> 0xC;
-            MoveEntity();
-            if (abs(slograGaibonDistX) < 8) {
-                if (abs(slograGaibonDistY) < 8) {
-                    self->ext.GS_Props.grabedAscending = 1;
-                    self->velocityX = 0;
-                    self->velocityY = 0;
-                    self->step_s++;
-                } else {
-                    if (SLOGRA.ext.GS_Props.pickupFlag == 0) {
-                        self->ext.GS_Props.grabedAscending = 0;
-                        SetStep(GAIBON_FLY_SHOOT_FIREBALLS);
-                    }
-                }
-            } else {
-                if (SLOGRA.ext.GS_Props.pickupFlag == 0) {
-                    self->ext.GS_Props.grabedAscending = 0;
-                    SetStep(GAIBON_FLY_SHOOT_FIREBALLS);
-                }
-            }
-            break;
-
-        case GAIBON_PICKUP_SLOGRA_ASCENDING:
-            AnimateEntity(D_801812AC, self);
-            if (self->animFrameIdx == 1 && self->animFrameDuration == 0) {
-                func_801C29B0(NA_SE_EN_GAIBON_FLAP_WINGS);
-            }
-            MoveEntity();
-            self->velocityY -= 0xA00;
-            if (self->velocityY < FIX(-2.0)) {
-                self->velocityY = FIX(-2);
-            }
-            self->ext.GS_Props.grabedAscending = 1;
-            SLOGRA.posX.i.hi = self->posX.i.hi;
-            SLOGRA.posY.i.hi = self->posY.i.hi + 28;
-            if (self->posY.i.hi < 16) {
-                self->velocityX = 0;
-                self->velocityY = 0;
-                self->ext.GS_Props.timer = 96;
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_PICKUP_SLOGRA_AIMING:
-            AnimateEntity(D_801812AC, self);
-            if (self->animFrameIdx == 1 && self->animFrameDuration == 0) {
-                func_801C29B0(NA_SE_EN_GAIBON_FLAP_WINGS);
-            }
-            if (GetSideToPlayer() & 1) {
-                self->velocityX -= 0xA00;
-            } else {
-                self->velocityX += 0xA00;
-            }
-            if (self->velocityX < FIX(-2.0)) {
-                self->velocityX = FIX(-2);
-            }
-            if (self->velocityX > FIX(2.0)) {
-                self->velocityX = FIX(2);
-            }
-            MoveEntity();
-
-            SLOGRA.velocityY = 0;
-            self->ext.GS_Props.grabedAscending = 0;
-            SLOGRA.posX.i.hi = self->posX.i.hi;
-            SLOGRA.posY.i.hi = self->posY.i.hi + 28;
-            self->ext.GS_Props.timer--;
-            if (self->ext.GS_Props.timer == 0) {
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_PICKUP_SLOGRA_RELEASE:
-            self->ext.GS_Props.grabedAscending = 0;
-            SetStep(GAIBON_FLY_TOWARDS_PLAYER);
-        }
-        break;
-
-    case GAIBON_NEAR_DEATH:
-        switch (self->step_s) {
-        case GAIBON_NEAR_DEATH_SETUP:
-            self->animCurFrame = 9;
-            self->velocityX = 0;
-            self->velocityY = 0;
-            self->step_s++;
-
-        case GAIBON_NEAR_DEATH_FLOOR_HIT_WAIT:
-            if (func_801BCCFC(&D_80181240) & 1) {
-                func_801C29B0(NA_SE_EN_GAIBON_SCREAM);
-                SetSubStep(GAIBON_NEAR_DEATH_FLOOR_LANDING);
-            }
-            break;
-
-        case GAIBON_NEAR_DEATH_FLOOR_LANDING:
-            if (AnimateEntity(D_801812F0, self) == 0) {
-                self->ext.GS_Props.flag = 0;
-                SetSubStep(GAIBON_NEAR_DEATH_TRANSFORM);
-                if (self->flags & FLAG_DEAD) {
-                    func_801C29B0(NA_SE_EN_GAIBON_COLLAPSE);
-                    SetStep(GAIBON_DYING);
-                } else {
-                    func_801C29B0(NA_SE_EN_GAIBON_SCREAM);
-                }
-            }
-            break;
-
-        case GAIBON_NEAR_DEATH_TRANSFORM:
-            if (AnimateEntity(D_801812FC, self) == 0) {
-                self->ext.GS_Props.flag++;
-                self->palette = self->ext.GS_Props.flag + D_80180D36;
-                if (self->ext.GS_Props.flag == 6) {
-                    self->flags &= ~0xF;
-                    SetStep(GAIBON_FLY_SHOOT_FIREBALLS);
-                }
-            }
-        }
-        break;
-
-    case GAIBON_DYING:
-        switch (self->step_s) {
-        case GAIBON_DYING_REACT:
-            animStatus = AnimateEntity(D_80181304, self);
-            if (animStatus == 0) {
-                self->ext.GS_Props.timer = 96;
-                self->animCurFrame = 31;
-                self->flags &= ~0xF;
-                do {
-                } while (0);
-                self->palette = D_80180D36;
-                func_801C29B0(NA_SE_EN_GAIBON_FLAME_OUT);
-                self->step_s++;
-            }
-            break;
-
-        case GAIBON_DYING_TURN_INTO_BONES:
-            if ((self->ext.GS_Props.timer % 8) == 0) {
-                func_801C29B0(NA_SE_EN_GAIBON_FLAME);
-                newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
-                if (newEntity != NULL) {
-                    CreateEntityFromEntity(E_FIRE, self, newEntity);
-                    newEntity->posY.i.hi += 28;
-                    newEntity->posX.i.hi -= 32 - (Random() & 63);
-                    newEntity->zPriority = self->zPriority + 1;
-                    newEntity->params = 0;
-                }
-            }
-            self->ext.GS_Props.timer--;
-            if (self->ext.GS_Props.timer == 0) {
-                g_BossFlag |= BOSS_FLAG_GAIBON_DEAD;
-                DestroyEntity(self);
-                return;
-            }
-        }
-        break;
-
-    case GAIBON_DEBUG:
-        FntPrint("charal %x\n", self->animCurFrame);
-        if (g_pads[1].pressed & PAD_SQUARE) {
-            if (self->params == 0) {
-                self->animCurFrame++;
-                self->params |= 1;
-            } else {
-                break;
-            }
-        } else {
-            self->params = 0;
-        }
-        if (g_pads[1].pressed & PAD_CIRCLE) {
-            if (self->step_s == 0) {
-                self->animCurFrame--;
-                self->step_s |= 1;
-            }
-        } else {
-            self->step_s = 0;
-        }
-        break;
-    }
-
-    if (!(self->flags & FLAG_DEAD)) {
-        slograGaibonDistX = self->posX.i.hi + g_Tilemap.scrollX.i.hi;
-        slograGaibonDistY = self->posY.i.hi + g_Tilemap.scrollY.i.hi;
-
-        if (self->velocityX > 0 && slograGaibonDistX > 784) {
-            self->posX.i.hi = 784 - g_Tilemap.scrollX.i.hi;
-        }
-
-        if (self->velocityX < 0 && slograGaibonDistX < 96) {
-            self->posX.i.hi = 96 - g_Tilemap.scrollX.i.hi;
-        }
-        if (self->velocityY > 0 && slograGaibonDistY > 420) {
-            self->posY.i.hi = 420 - g_Tilemap.scrollY.i.hi;
-        }
-
-        hitbox = &D_80181340[self->animCurFrame][D_80181310];
-        hitbox--;
-        hitbox++;
-        self->unk10 = *hitbox++;
-        self->unk12 = *hitbox++;
-        self->hitboxWidth = hitbox[0];
-        self->hitboxHeight = hitbox[1];
-    }
-}
-#endif
-
-void func_801B69E8(Entity* self) {
-    Entity* prevEntity;
-    s16 animCurFrame;
-
-    if (self->step == 0) {
-        InitializeEntity(D_80180D30);
-        self->hitboxState = 0;
-    }
-
-    prevEntity = &self[-1];
-    self->facingLeft = prevEntity->facingLeft;
-    self->palette = prevEntity->palette;
-    self->posX.i.hi = prevEntity->posX.i.hi;
-    self->posY.i.hi = prevEntity->posY.i.hi;
-    self->animCurFrame = 0;
-
-    if ((prevEntity->animCurFrame - 32) < 3U) {
-        self->animCurFrame = 0x26;
-    } else if (prevEntity->animCurFrame == 35) {
-        self->animCurFrame = 0x27;
-    } else if ((prevEntity->animCurFrame - 36) < 2U) {
-        self->animCurFrame = 0x28;
-    }
-
-    if (prevEntity->entityId != E_GAIBON) {
-        DestroyEntity(self);
-    }
-}
-
-// small red projectile from gaibon
-void EntitySmallGaibonProjectile(Entity* self) {
-    if (self->flags & FLAG_DEAD) {
-        self->pfnUpdate = EntityExplosion;
-        self->drawFlags = 0;
-        self->step = 0;
-        self->entityId = 2;
-        self->params = 0;
-        return;
-    }
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180D3C);
-        self->animSet = ANIMSET_DRA(2);
-        self->animCurFrame = 1;
-        self->drawFlags = FLAG_DRAW_ROTX | FLAG_DRAW_ROTZ;
-        self->rotX = 0xC0;
-        self->velocityX = (rcos(self->rotZ) * 0x28000) >> 0xC;
-        self->velocityY = (rsin(self->rotZ) * 0x28000) >> 0xC;
-        self->palette = 0x81B6;
-        self->rotZ -= 0x400;
-
-    case 1:
-        MoveEntity();
-        AnimateEntity(D_8018136C, self);
-        break;
-    }
-}
-
-// large red projectile from gaibon
-void EntityLargeGaibonProjectile(Entity* self) {
-    Entity* newEntity;
-
-    if (self->flags & FLAG_DEAD) {
-        self->pfnUpdate = EntityExplosion;
-        self->entityId = 2;
-        self->drawFlags = 0;
-        self->step = 0;
-        self->params = 1;
-        return;
-    }
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180D48);
-        if (self->params == 0) {
-            self->animSet = ANIMSET_DRA(2);
-            self->drawFlags = FLAG_DRAW_ROTZ;
-            self->velocityX = (rcos(self->rotZ) * 0x38000) >> 0xC;
-            self->velocityY = (rsin(self->rotZ) * 0x38000) >> 0xC;
-            self->palette = 0x81B6;
-            self->rotZ -= 0x400;
-        } else {
-            self->animSet = ANIMSET_DRA(14);
-            self->unk5A = 0x79;
-            self->drawFlags = FLAG_DRAW_ROTX | FLAG_DRAW_ROTZ | FLAG_DRAW_UNK8;
-            self->rotX = 0x100;
-            self->unk6C = 0x80;
-            self->palette = 0x81F3;
-            self->drawMode = 0x30;
-            self->step = 2;
-            self->hitboxState = 0;
-            self->flags |= FLAG_UNK_2000;
-        }
-        break;
-
-    case 1:
-        MoveEntity();
-        AnimateEntity(D_80181378, self);
-        if (!(g_Timer & 3)) {
-            newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
-            if (newEntity != NULL) {
-                CreateEntityFromEntity(E_GAIBON_BIG_FIREBALL, self, newEntity);
-                newEntity->params = 1;
-                newEntity->rotZ = self->rotZ;
-                newEntity->zPriority = self->zPriority + 1;
-            }
-        }
-        break;
-
-    case 2:
-        self->unk6C += 0xFE;
-        self->rotX -= 4;
-        if (AnimateEntity(D_80181388, self) == 0) {
-            DestroyEntity(self);
-        }
         break;
     }
 }
